completed implementation of sea ice thickness
