+++ test/mock_results/test_sea_ice_cryosat-2/request.json
@@ -1,0 +1,1 @@
+{"_dataset_name": "satellite-sea-ice-thickness", "satellite": "cryosat-2", "cdr_type": "cdr", "version": "2_0", "variable": "all", "format": "tgz", "month": ["03", "04"], "year": "2016"}
